switch from `scripts` to `console_scripts` entry
`console_scripts` allows for python functions from a package to be used
as an entry point for an application. the `scripts` entry allows for
anything to be used, such as a shell script.

as the script is python anyway, using `console_scripts` is ok. it also
has the added benefit of making it windows compatible.